core/requirements.txt
Django==5.2.1
django-modeltranslation==0.18.12
Pillow==10.1.0 